prompt1
Create a modern landing page for a organization called "VOTUM" supporting people with disabilities. The style should be friendly, positive, inclusive, easy to navigate. Use soft colors as light blue, pastel tones. The page should be responsive, accessible and adhere to WCAG guidlines. The page stucture should include a header with logo, name of the organization,  buttons to increase and decrease font size and a radio button to change language from slovak to english. The second row of header should consist of navigation menu with icons and labels under them named: Home, About us, Events, History, Support, Contacts and Documents. Under header there should be a hero section with slogan, buttons like  “more about us”,Facebook and youtube button and a group picture of young people. Under that, upcoming events section devided into two parts - list of events on the right side, a calendar on left side with these events marked and a button to "add to your calendar" under. Next section are few featured events with a button to see more and on the bottom a picture of our team and a button to "learning more about our team". I need this project to be written in PHP, PHP Blade using Laravel framework and Tailwind styling. Use the wireframes i sent you as a layout you should stick to.