prompt2
lets work on the header. I would like the option of the font size change to be in a format of "- font +" and the locale buttons to have the flags of slovakia and england above SK and EN. Also, make the whole header more visually appealing and edit the radio buttons to look like bootstrap's seamless looking button group but remove any useless parts of code.