prompt4
lets get back to the header. I would like the language to have a globe icon and a SK/EN option. I dont like the design of the font change and the language change. Please try to make it less boxy and more natural, you can also change the colours of the header and navbar to be in blue tones but not too dark