prompt6
can you make the header fixed at the top but not full width, more of a modern, minimalistic „sticky top navigation header“, the hero section's background would go to the top and the header would stay on the top even after scrolling. I also dont like the black line between header and navbar, remove it please or make it less visible. remove all the borders and shadows from navbar. you will need to add a margin on top of hero section so it is not covered by header but the headers size will probably change depending on font size set