prompt7
opening of menu on phone stopped working, can you repair the toggle? when i click the menu button i can see a few pixel box on the left side but the menu is not visible, i dont know what could be the problem it worked before making the header fixed on top.....is there any other way? i want the navbar to be a part of the floating bubble and also i want it to be a bubble not full width, can it for example open a offcanvas drawer like in bootstrap?