ChatGPT prompt2
Lets work on header. Make the lower part of the header - the part with navigation just a little bit lighter than the top part, but still dark blue. Edit the font size and language buttons so they look more modern and fit more into the color theme of the header. Make the header responsive. On the mobile phone the buttons and the links should hide behind a menu button. When the menu button is clicked, a full width menu should open under the header. It should consist of font size and language options in one row and then every row should be for one navigation link with icon.