ChatGPT prompt6
Let's move on to the next page - history, which will appear after clicking the history navigation button. This page should contain a title - History of Združenie Votum. The rest of the page will consist of a vertical timeline, always with a year and text, some years may also have one picture. Fill this timeline with at least 12 made-up historical events. Put a home button with a logo of home under this timeline and link it to the landing page you created before