ChatGPT prompt 8
Let’s work on the page called AboutUs. The page should be divided into two parts - information about our organization - our goals, vision, mission etc., each of these in its own sections with title, text and a photo or a pictogram. The second part of the page will contain information about our team with a photo of a member, name and a short text about each of the members. The top of the page should also contain a button to listen to the text and a button to share this page. There should also be a home button on the bottom of the page like on the history page. Take inspiration from the wireframe I sent you.