ChatGPT prompt9
Let’s work on the Events page. The whole page should consist of events with a picture, title and a one sentence description. On the top of the page, there should be visually separated future events. Under that the rest of the events should be in chronological order from the newest. There should be an option to filter by year on the top and also an indicated pager on the bottom.  There should also be a home button on the bottom of the page like on all of the pages. Take inspiration from the wireframe I sent you.