ChatGPT prompt10
Now let’s generate the page of the single event when a person clicks on any of the events. On the top, there should  be a date and the same buttons from the about us page - listening and sharing. Under that will be a section of text, next a section for pictures, videos and the last is a section for documents. On the bottom, as always a home button and a button to the previous page - events. Take inspiration from the wireframe I sent you.