ChatGPT prompt13
I need to generate the page that will open after clicking on the financial support option on the support page. It will contain a guide on each option of financial support. That is - via QR code or via bank transfer. This page should especially be well aranged, easy to read and understand with all important information ready to use. Take inspiration from the wireframe I sent you.