ChatGPT prompt15
The last type of support and the last page is - other types of support. This should contain options like offers for musical performances (because we have our own band and we do music therapy), fundraiser, and donations. I did not prepare a wireframe for this page so be creative and come up with something suitable for this page.